RF: Consolidate all utils actually unique to `get` (plain move)
